SOLR-12348: move start collection creation logic to TimeRoutedAlias
